Adding the Detail Activity

+++ data/src/main/java/com/sebasorozcob/www/data/remote/dto/CreditDto.kt
@@ -8,7 +8,7 @@ data class CreditDto(
     @SerializedName("id") val id: Int,
     @SerializedName("known_for_department") val knownDepartment: String,
     @SerializedName("name") val name: String,
-    @SerializedName("popularity") val popularity: Int,
+    @SerializedName("popularity") val popularity: Double,
     @SerializedName("profile_path") val profilePhoto: String,
     @SerializedName("cast_id") val castId: Int,
     @SerializedName("character") val character: String,
